+++ lib/signalwire/relay/calling/action/fax_action.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require 'forwardable'
+
+module Signalwire::Relay::Calling
+  class FaxAction < Action
+    def result
+      FaxResult.new(@component)
+    end
+
+    def stop
+      @component.stop
+    end
+  end
+end