+++ lib/signalwire/relay/calling/result/fax_result.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Signalwire::Relay::Calling
+  class FaxResult < Result
+    def_delegators :@component, :direction, :identity, :remote_identity, :document, :pages
+  end
+end