+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `:task` broadcast from client
+- `Relay::Task` and `on_task` handler for `Consumer`
+
 ## [2.0.0] - 2019-07-16
 ### Added
 - Connection Retry upon disconnect.