+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Fax API for Relay
+
 ## [2.0.0] - 2019-07-16
 ### Added
 - Connection Retry upon disconnect.